vdoke/requirements.txt
python-vlc>=3.0.18
pygame>=1.9.5
pafy
moviepy
youtube_dl